% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{axis}
\alias{axis}
\alias{mobile_hide_axis}
\alias{mobile_axis_x}
\alias{mobile_axis_y}
\title{Axis}
\usage{
mobile_hide_axis(m)

mobile_axis_x(m, ...)

mobile_axis_y(m, ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/coordinate}{official documentation}.}
}
\description{
Customise axis.
}
\examples{
df <- data.frame(
 label = c("banana", "apple", "cake"),
 value = c(50, 30, 20),
 x = "1" # phoney variable
)

mobile(df, aes(x, value, color = label, adjust = stack)) \%>\% 
 mobile_bar() \%>\% 
 mobile_coord("polar") \%>\% 
 mobile_hide_axis()

}
