% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom}
\alias{geom}
\alias{mobile_bar}
\alias{mobile_point}
\alias{mobile_path}
\alias{mobile_line}
\alias{mobile_area}
\alias{mobile_polygon}
\alias{mobile_schema}
\title{Geoms}
\usage{
mobile_bar(m, ..., data = NULL, inherit_aes = TRUE)

mobile_point(m, ..., data = NULL, inherit_aes = TRUE)

mobile_path(m, ..., data = NULL, inherit_aes = TRUE)

mobile_line(m, ..., data = NULL, inherit_aes = TRUE)

mobile_area(m, ..., data = NULL, inherit_aes = TRUE)

mobile_polygon(m, ..., data = NULL, inherit_aes = TRUE)

mobile_schema(m, ..., data = NULL, inherit_aes = TRUE)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{...}{Aesthetics.}

\item{data}{A data.frame.}

\item{inherit_aes}{Whether to inherit the aesthetics from \code{\link{mobile}}.}
}
\description{
Add a geometry.
}
\examples{
mobile(cars) \%>\% 
 mobile_point(aes(speed, dist))

}
