% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile.R
\name{mobile}
\alias{mobile}
\title{Initialise}
\usage{
mobile(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{...}{Mobile chart aesthetics.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{ID of element holding the mobile chart.}
}
\description{
Initialise a chart.
}
\examples{
mobile(cars) \%>\% 
 mobile_point(aes(speed, dist))

}
