% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{mobile_animate}
\alias{mobile_animate}
\title{Animate}
\usage{
mobile_animate(m, aspect, ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{aspect}{Aspect to animate, see section below for valid values.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/coordinate}{official documentation}.}
}
\description{
Animate various aspects of the chart.
}
\section{Aspect}{

Valid \code{aspect}:
\itemize{
 \item{\code{axis-label}}
 \item{\code{axis-grid}}
 \item{\code{axis-tick}}
 \item{\code{axis-line}}
 \item{\code{line}}
 \item{\code{area}}
 \item{\code{bar}}
 \item{\code{path}}
 \item{\code{point}}
 \item{\code{polygon}}
 \item{\code{schema}}
}
}

\examples{
mobile(cars, aes(speed, dist)) \%>\% 
 mobile_point() \%>\% 
 mobile_animate(
   "point", 
   appear = list(
     animation = 'fadeIn',
     duration = 5000
   )
 )

}
