% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{mobile_legend}
\alias{mobile_legend}
\title{Legend}
\usage{
mobile_legend(m, ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/legend}{official documentation}.}
}
\description{
Customise the legend of the chart.
}
\examples{
mobile(cars) \%>\% 
 mobile_point(aes(speed, dist)) \%>\% 
 mobile_legend(position = "right")

}
