% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{mobile_options}
\alias{mobile_options}
\title{General Options}
\usage{
mobile_options(m, padding = "auto", animate = TRUE)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{padding}{An integer or a list containing 4 integers
e.g.: \code{list(40, 10, 'auto', 'auto')} giving padding
on top, right, bottom, and left.}

\item{animate}{Set to \code{FALSE} to disable animations.}
}
\description{
Pass general options to the chart.
}
