% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll.R
\name{mobile_scroll}
\alias{mobile_scroll}
\title{Scroll}
\usage{
mobile_scroll(m, mode = c("x", "y", "xy"), ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{mode}{Axis to apply scroll to.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/scroll-bar}{official documentation}.}
}
\description{
Add a scrollbar to let user pan charts.
}
\examples{
# range for pan
lmt <- list(x = list(min = -20, max = 20))

mobile(df, aes(x, y)) \%>\% 
 mobile_bar() \%>\% 
 mobile_interaction("pan", limitRange = lmt) \%>\% 
 mobile_scroll(mode = "x", xStyle = list(offsetY = -5))

}
