% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{mobile_tooltip}
\alias{mobile_tooltip}
\title{Tooltip}
\usage{
mobile_tooltip(m, ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/tooltip}{official documentation}.}
}
\description{
Add a tooltip. Note that the visualisations being optimized for mobile there
is no such thing as hover: tooltips show on click.
}
\examples{
mobile(cars) \%>\% 
 mobile_point(aes(speed, dist)) \%>\% 
 mobile_tooltip(snap = TRUE)

}
