% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\alias{mobile_scale_x}
\alias{mobile_scale_y}
\title{Scale}
\usage{
mobile_scale_x(m, ...)

mobile_scale_y(m, ...)
}
\arguments{
\item{m}{An object of class \code{mobile} as returned by \code{\link{mobile}}.}

\item{...}{Named options from the 
\href{https://f2.antv.vision/en/docs/api/chart/scale}{official documentation}.}
}
\description{
Scale variables.
}
\examples{
mobile(cars, aes(speed, dist)) \%>\% 
 mobile_point() \%>\% 
 mobile_scale_x(min = -10, tickCount = 20)

}
